requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
concurrent-log-handler==0.9.12
ConcurrentLogHandler==0.9.1
cryptography==2.3.1
dj-stripe==1.2.3
Django==1.11.14
django-rest-framework==0.1.0
djangorestframework==3.8.2
enum34==1.1.6
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
jsonfield==2.0.2
pkg-resources==0.0.0
pycparser==2.19
pyOpenSSL==18.0.0
pytz==2018.5
requests==2.20.1
six==1.11.0
stripe==2.12.0
urllib3==1.24.1
